\name{event}
\title{Event History Analysis Library}
\description{
\code{\link{bp}}			Create a Vector of Cumulative Numbers of Previous Events for a Point Process

\code{\link{coxre}}		Cox Proportional Hazards Model with Random Effect

\code{\link{cprocess}}	Counting Process Plot

\code{\link{ehr}}		Fit an Intensity Function to Event Histories

\code{\link{hboxcox}}		Log Hazard Function for a Box-Cox Process

\code{\link{hburr}}		Log Hazard Function for a Burr Process

\code{\link{hcauchy}}		Log Hazard Function for a Cauchy Process

\code{\link{hexp}}		Log Hazard Function for an Exponential (Poisson) Process

\code{\link{hgamma}}		Log Hazard Function for a Gamma Process

\code{\link{hgextval}}	Log Hazard Function for an Extreme Value  Process

\code{\link{hggamma}}		Log Hazard Function for a Generalized Gamma Process

\code{\link{hglogis}}		Log Hazard Function for a Generalized Logistic Process

\code{\link{hgweibull}}	Log Hazard Function for a Generalized Weibull Process

\code{\link{hhjorth}}		Log Hazard Function for a Hjorth Process

\code{\link{hinvgauss}}	Log Hazard Function for a Inverse Gauss Process

\code{\link{hlaplace}}	Log Hazard Function for a Laplace Process

\code{\link{hlnorm}}		Log Hazard Function for a Log Normal Process

\code{\link{hlogis}}		Log Hazard Function for a Logistic Process

\code{\link{hnorm}}		Log Hazard Function for a Normal Process

\code{\link{hpareto}}		Log Hazard Function for a Pareto Process

\code{\link{hstudent}}	Log Hazard Function for a Student Process

\code{\link{hweibull}}	Log Hazard Function for a Weibull Process

\code{\link{ident}}		Create an Individual Identification Vector for a Point Process

\code{\link{kalsurv}}		Generalized Repeated Measurements Models for Event Histories

\code{\link{km}}			Kaplan-Meier Survival Curves

\code{\link{pbirth}}		Fit Overdispersed Count Data as a Birth Process

\code{\link{pp}}			Create a Point Process Vector from Times between Events

\code{\link{read.list}}	Read a List of Matrices of Unbalanced Repeated Measurements from a File

\code{\link{read.surv}}	Read a List of Vectors of Event Histories from a File

\code{\link{survkit}}		Weibull and Cox Models with Random Effects

\code{\link{tccov}}		Create a Vector of Time-constant Covariates for a Point Process

\code{\link{tpast}}		Create a Vector of Times Past since Previous Events for a Point Process

\code{\link{ttime}}		Create a Vector of Total Time Elapsed for each Individual for a Point Process

\code{\link{tvcov}}		Create a Vector of Time-varying Covariates for a Point Process
}
