\name{event}
\title{Event History Analysis Library}
\alias{event}
\description{
\code{\link[event]{autointensity}}	Plot Autointensity Function of a Point Process

\code{\link[event]{bp}}			Create a Vector of Cumulative Numbers of Previous Events for a Point Process

\code{\link[event]{coxre}}		Cox Proportional Hazards Model with Random Effect

\code{\link[event]{cprocess}}	Counting Process Plot

\code{\link[event]{ehr}}		Fit an Intensity Function to Event Histories

\code{\link[event]{hboxcox}}		Log Hazard Function for a Box-Cox Process

\code{\link[event]{hburr}}		Log Hazard Function for a Burr Process

\code{\link[event]{hcauchy}}		Log Hazard Function for a Cauchy Process

\code{\link[event]{hexp}}		Log Hazard Function for an Exponential (Poisson) Process

\code{\link[event]{hgamma}}		Log Hazard Function for a Gamma Process

\code{\link[event]{hgextval}}	Log Hazard Function for an Extreme Value  Process

\code{\link[event]{hggamma}}		Log Hazard Function for a Generalized Gamma Process

\code{\link[event]{hglogis}}		Log Hazard Function for a Generalized Logistic Process

\code{\link[event]{hgweibull}}	Log Hazard Function for a Generalized Weibull Process

\code{\link[event]{hhjorth}}		Log Hazard Function for a Hjorth Process

\code{\link[event]{hinvgauss}}	Log Hazard Function for a Inverse Gauss Process

\code{\link[event]{hlaplace}}	Log Hazard Function for a Laplace Process

\code{\link[event]{hlnorm}}		Log Hazard Function for a Log Normal Process

\code{\link[event]{hlogis}}		Log Hazard Function for a Logistic Process

\code{\link[event]{hnorm}}		Log Hazard Function for a Normal Process

\code{\link[event]{hpareto}}		Log Hazard Function for a Pareto Process

\code{\link[event]{hskewlaplace}}	Log Hazard Function for a Skew Laplace Process

\code{\link[event]{hstudent}}	Log Hazard Function for a Student Process

\code{\link[event]{hweibull}}	Log Hazard Function for a Weibull Process

\code{\link[event]{ident}}		Create an Individual Identification Vector for a Point Process

\code{\link[event]{kalsurv}}		Generalized Repeated Measurements Models for Event Histories

\code{\link[event]{km}}			Kaplan-Meier Survival Curves

\code{\link[event]{pbirth}}		Fit Overdispersed Count Data as a Birth Process

\code{\link[event]{pp}}			Create a Point Process Vector from Times between Events

\code{\link[rmutil]{read.list}}	Read a List of Matrices of Unbalanced Repeated Measurements from a File

\code{\link[rmutil]{read.surv}}	Read a List of Vectors of Event Histories from a File

\code{\link[event]{survkit}}		Weibull and Cox Models with Random Effects

\code{\link[event]{tccov}}		Create a Vector of Time-constant Covariates for a Point Process

\code{\link[event]{tpast}}		Create a Vector of Times Past since Previous Events for a Point Process

\code{\link[event]{ttime}}		Create a Vector of Total Time Elapsed for each Individual for a Point Process

\code{\link[event]{tvcov}}		Create a Vector of Time-varying Covariates for a Point Process
}
\keyword{documentation}
