% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.r
\name{hstudent}
\alias{hstudent}
\title{Hazard Function for a Student t Process}
\usage{
hstudent(y, m, s, f)
}
\arguments{
\item{y}{Vector of times.}

\item{m}{Location parameter.}

\item{s}{Scale parameter.}

\item{f}{Degrees of freedom.}
}
\value{
\code{hstudent} returns the log hazard function for a Student t
process with the given parameter values.
}
\description{
Hazard Function for a Student t Process
}
\examples{

hstudent(1:10, 3, 2, 5)

}
\author{
J.K. Lindsey
}
\seealso{
\code{\link[rmutil]{hboxcox}}, \code{\link[rmutil]{hburr}},
\code{\link[event]{hexp}}, \code{\link[rmutil]{hgextval}},
\code{\link[event]{hcauchy}}, \code{\link[event]{hgamma}},
\code{\link[rmutil]{hggamma}}, \code{\link[rmutil]{hhjorth}},
\code{\link[rmutil]{hinvgauss}}, \code{\link[rmutil]{hlaplace}},
\code{\link[event]{hlogis}}, \code{\link[rmutil]{hglogis}},
\code{\link[event]{hnorm}}, \code{\link[event]{hnorm}},
\code{\link[event]{hweibull}}, \code{\link[rmutil]{hgweibull}}.
}
\keyword{distribution}

