% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr.r
\name{ident}
\alias{ident}
\title{Create an Individual Identification Vector for a Point Process}
\usage{
ident(y, id)
}
\arguments{
\item{y}{Vector of times.}

\item{id}{Vector of corresponding individual identifiers for who had which
sequence of times.}
}
\value{
\code{ident} creates a vector of length \code{sum(y)} by repeating
the values of individual identifiers for the times for use with
\code{\link[event]{ehr}}.
}
\description{
Create an Individual Identification Vector for a Point Process
}
\examples{

y <- c(5,3,2,4)
i <- c(1,1,2,2)
id <- ident(y, i)
id

}
\author{
J.K. Lindsey
}
\seealso{
\code{\link[event]{bp}}, \code{\link[event]{ehr}},
\code{\link[event]{pp}}, \code{\link[event]{tccov}},
\code{\link[event]{tpast}}, \code{\link[event]{ttime}},
\code{\link[event]{tvcov}}.
}
\keyword{manip}

