% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbirth.r
\name{pbirth}
\alias{pbirth}
\title{Fit Overdispersed Count Data as a Birth Process}
\usage{
pbirth(frequencies, p, intensity = "negative binomial",
  type = "spectral decomposition", print.level = 0, ndigit = 10,
  gradtol = 1e-05, steptol = 1e-05, fscale = 1, iterlim = 100,
  typsize = abs(p), stepmax = 10 * sqrt(p \%*\% p))
}
\arguments{
\item{frequencies}{Vector of frequencies or a matrix with each row a
different series of frequencies.}

\item{p}{Vector of initial estimates.}

\item{intensity}{The intensity function of the process: binomial, binomial
exdponential, binomial logistic, binomial total, Poisson, Poisson
exponential, negative binomial, or gen(eralized) negative binomial.}

\item{type}{Algorithm used for matrix exponentiation: spectral decomposition
or series approximation.}

\item{others}{Arguments controlling \code{\link{nlm}}.}
}
\description{
\code{pbirth} fits binomial, binomial exponential, binomial logistic,
binomial total, Poisson, Poisson exponential, negative binomial,
gen(eralized) negative binomial, and generalized negative binomial processes
as a birth process.
}
\examples{

y <- rnbinom(100,2,0.6)
fr <- tabulate(y)
pbirth(fr, p=log(-log(0.7)), intensity="Poisson", type="series")
pbirth(fr, p=c(log(-log(0.7)),log(5)),
	intensity="negative binomial", type="series")
pbirth(fr, p=c(log(-log(0.7)),log(5),-1),
	intensity="gen negative binomial", type="series")

}
\author{
J.K. Lindsey
}
\references{
Faddy, M.J. and Fenlon, J.S. (1999) Stochastic modelling of the
invasion process of nematodes in fly larvae. Applied Statistics 48: 31-37.
}
\keyword{models}

