% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr.r
\name{tpast}
\alias{tpast}
\title{Create a Vector of Times Past since Previous Events for a Point Process}
\usage{
tpast(y)
}
\arguments{
\item{y}{Vector of times.}
}
\value{
\code{tpast} creates a vector of length \code{sum(y)} of times since
the immediately preceding event occurred for use with
\code{\link[event]{ehr}}.
}
\description{
Create a Vector of Times Past since Previous Events for a Point Process
}
\examples{

y <- c(5,3,2,4)
ptime <- tpast(y)
ptime

}
\author{
J.K. Lindsey
}
\seealso{
\code{\link[event]{bp}}, \code{\link[event]{ehr}},
\code{\link[event]{ident}}, \code{\link[event]{pp}},
\code{\link[event]{tccov}}, \code{\link[event]{ttime}},
\code{\link[event]{tvcov}}.
}
\keyword{manip}

